requirements.txt
django
django-ninja
django-cors-headers
pillow
yookassa
gunicorn
psycopg2-binary
python-dotenv 